Extracted VelJoint class to seperate code and header files.
